Bumped scripts to typescript
Also removed gulp because getting it to work with typescript is way too painful
