--- gulpfile.esm.js/index.js
@@ -1,5 +1,0 @@
-export { default as build } from './build';
-export { default as buildBundles } from './build/bundles';
-export { default as buildTabs } from './build/tabs';
-export { default as buildDocs } from './build/docs';
-export { default as create } from './templates';
