+++ scripts/templates/templates/__bundle__.ts
