+++ scripts/templates/templates/__tab__.tsx
